Escape hyphen from Lua patterns
Hyphen is a pretty common character to use in URLs to require
workarounds to enter, and the - feature is little used in
Lua patterns, so I'm choosing one over the other.
I'm not escaping the other characters, in case someone is using
them. They are not officially supported and I'm not sure it's
a good idea to keep them, but I'm leaving this underspecified
for now.

Fixes #2.